aruha-563: added support for ',' before brackets
